Refactored y2023::d9.
Using a Grid2D instead of a dynamically-sized trapezoidal array is a lot
simpler and easier to grok. I also restructured Solution to have one
vector of ranges and one of values to reduce the necessary held
allocations. The extrapolations are computed just after parsing
internally, before the standard parse returns, using a single grid
that's allocated once for all Histories.